+++ examples/downloader.py
@@ -1,7 +1,5 @@
 """
-
 A rudimentary URL downloader (like wget or curl) to demonstrate Rich progress bars.
-
 """
 
 from concurrent.futures import ThreadPoolExecutor