+++ examples/highlighter.py
@@ -1,3 +1,7 @@
+"""
+This example demonstrates a simple text highlighter.
+"""
+
 from rich.console import Console
 from rich.highlighter import RegexHighlighter
 from rich.theme import Theme